backend/requirements.txt
# Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# AI/ML
sentence-transformers==2.3.1
faiss-cpu==1.13.2
google-generativeai==0.3.2
openai==1.10.0

# Firebase
firebase-admin==6.3.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
PyPDF2==3.0.1
